requirements.txt
# Python dependencies for scripts/ (oauth_setup.py, load_accounts.py)
requests>=2.28
psycopg2-binary>=2.9
